Initial location for libformat for coloring common high level types.
Currently:
	- IRI's
	- Timestamps and Dates
	- Local Files

While appropriately associated with terminal, some of the functionality should be generalized
by tokenizing the parts of the data type in order to allow convenient formatting. Notably,
the IRI/URI code should be relocated so that it can properly escape the fields in the tokens
so that they may be simply catenated.